requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-groq==0.1.5
langchain-openai==0.1.20
langchain-core==0.2.32
langchain-cohere==0.1.8
langchainhub==0.1.14
langchain-experimental==0.0.64
langsmith==0.1.93
rank-bm25==0.2.2
chromadb==0.5.5
fuzzywuzzy==0.18.0
elasticsearch==8.15.0
openpyxl==3.1.5
gradio==4.39.0
streamlit==1.37.0
python-dotenv==1.0.1
faiss-cpu==1.8.0
google-api-python-client>=2.100.0
beautifulsoup4
html2text